pyproject.toml
[project]
name = "hackathon-adk-on-gcp"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "flet[all]==0.28.3",
    "google-adk>=1.11.0",
    "google-cloud-aiplatform==1.110.0",
    "ipython>=9.4.0",
    "langchain==0.3.27",
    "langchain-community>=0.3.27",
    "langchain-core==0.3.74",
    "langchain-google-vertexai==2.0.28",
    "langchain-openai>=0.3.30",
    "langchain-tavily==0.2.11",
    "langgraph==0.3.34",
    "langsmith==0.4.15",
    "openinference-instrumentation-langchain==0.1.34",
    "pillow>=11.3.0",
    "python-dotenv==1.1.1",
    "typing-extensions==4.14.1",
]
